editoast: create zone.rs in primitives folder
